refactor: move code into central file
