+++ tile/assemble_test.go
